feat: use shared app shell on profile pages
Move profile under the authenticated shell so settings keep the home sidebar chrome.
