+++ src/components/layouts/app-sidebar.tsx
@@ -33,6 +33,7 @@ export function AppSidebar({ onNavigate }: AppSidebarProps) {
   const t = useTranslations("navigation");
   const { data: session } = useSession();
   const user = session?.user;
+  const isProfileActive = pathname === "/profile" || pathname?.startsWith("/profile/");
 
   return (
     <aside className="flex h-full w-64 shrink-0 flex-col border-r border-sidebar-border bg-sidebar text-sidebar-foreground">
@@ -76,7 +77,12 @@ export function AppSidebar({ onNavigate }: AppSidebarProps) {
         <Link
           href="/profile"
           onClick={onNavigate}
-          className="flex items-center gap-3 rounded-lg px-2 py-2 text-sm font-medium text-muted-foreground transition-colors hover:bg-sidebar-accent hover:text-foreground"
+          className={cn(
+            "flex items-center gap-3 rounded-lg px-2 py-2 text-sm font-medium transition-colors",
+            isProfileActive
+              ? "bg-accent/15 text-accent"
+              : "text-muted-foreground hover:bg-sidebar-accent hover:text-foreground"
+          )}
         >
           <Avatar className="size-8">
             <AvatarImage src={user?.image} alt={user?.name || "Profile"} />
